+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Breaking Change
 - mysql_class.Server.connect:  Moved change database to within the connection string.
 
 ### Added
+- mysql_class.MasterRep.repconn:  Setups a connection to a replication server.
 - mysql_class.Server.is_connected:  Checks to see if the connection is still active.
 - mysql_class.Server.reconnect:  Reconnects to database if connect is non-active.
 - mysql_class.Server.chg_db:  Change to another database.